Tutorial level updated
Imported the new version of the scene from my other project for the tutorial level into this project and deleted the old version of the tutorial level in this project.

New version has extended the rows of cubes (seats) to have more of them and made the train car dimensions bigger in general. Applied grey and yellow materials to make everything more visible.