Wires Finalization
Changed the final solution and have remapped the colours in line with the original design, no orange unfortunately but I used what was available as I'm not adding extra textures as this might break something else.

Implemented the visual lights so that when the puzzle is turned on by a button a red light shows indicating it is on and a green light for when the puzzle is correct. These will turn off if the next puzzle along is failed.