janitor door mesh renderer toggles with puzzle state
Once padlock puzzle is complete the mesh renderer for the janitor door becomes false, once the code is in the wrong order the mesh renderer comes back on. This should be temporary until the door animations are done.

I have also added a puzzles folder and duplicated the latest hospital scene in my folder.