Added Tooltips
Added in tooltips to scene, had to duplicate the tooltip script and edit it slightly to add a bool that checks when the input is completed so that the tooltip doesn't re-activate when you go back in range.

Also added tooltips and crouch detection to their own layers as it was causing issues.