Fixed jumping on pause menu and added prompt for save delete
I have fixed the bug where you could spam the jump button on a menu in-game, and then fly into the void on resuming, which I fixed by disabling the jump script when any of the menus are opened.

I also added the prompts for when you go to delete a save file as well, and there's an individual prompt for each delete button, so that a script can be added to the yes button which deletes that specific save file.